Added comparing text before and after its async checking.
Checking a large part of text can take an unknown amount of time.
So we have to compare the text before and after async work.
If the text has changed during async and we have more async,
we don't perform further refreshing of cache and underlines.
But if it was the last async, we should invoke a new one.