enchant: allow toggling individual dictionaries
Allow disabling and enabling individual dictionaries with the
Enchant/Hunspell backend.

This is mostly relevant in shared systems (where different users want to
enable different sets of languages) or on distributions which install an
exorbitant amount of dictionaries by default.

The SupportsToggleDictionaries() method should be used to determine
whether this is supported in the current environment. The macOS and
Windows both currently return false, with the latter apparently being
feasible but out of scope.